+++ class-teachingplatform/src/main/java/com/school/dao/inter/teachpaltform/paper/IStuPaperLogsDAO.java
@@ -7,7 +7,7 @@ import com.school.entity.teachpaltform.paper.StuPaperLogs;
 import java.util.*;
 
 public interface IStuPaperLogsDAO extends ICommonDAO<StuPaperLogs>{
-    public List<StuPaperLogs> getMarkingLogs(Integer paperid,Integer quesid);
-    public List<Map<String,Object>> getMarkingDetail(Integer paperid,Integer quesid);
-    public List<Map<String,Object>> getMarkingNum(Integer paperid,Integer quesid);
+    public List<StuPaperLogs> getMarkingLogs(Long paperid,Long quesid);
+    public List<Map<String,Object>> getMarkingDetail(Long paperid,Long quesid);
+    public List<Map<String,Object>> getMarkingNum(Long paperid,Long quesid);
 }
