+++ class-teachingplatform/src/main/java/com/school/manager/inter/teachpaltform/paper/IStuPaperLogsManager.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Map;
 
 public interface IStuPaperLogsManager  extends IBaseManager<StuPaperLogs> {
-    public List<StuPaperLogs> getMarkingLogs(Integer paperid,Integer quesid);
-    public List<Map<String,Object>> getMarkingDetail(Integer paperid,Integer quesid);
-    public List<Map<String,Object>> getMarkingNum(Integer paperid,Integer quesid);
+    public List<StuPaperLogs> getMarkingLogs(Long paperid,Long quesid);
+    public List<Map<String,Object>> getMarkingDetail(Long paperid,Long quesid);
+    public List<Map<String,Object>> getMarkingNum(Long paperid,Long quesid);
 }
